Moved database import into the database crate
