+++ server/service/grpc/transaction_service.rs
@@ -18,6 +18,10 @@ use compiler::query_structure::QueryStructure;
 use concept::{thing::thing_manager::ThingManager, type_::type_manager::TypeManager};
 use database::{
     database_manager::DatabaseManager,
+    query::{
+        execute_schema_query, execute_write_query_in_schema, execute_write_query_in_write, StreamQueryOutputDescriptor,
+        WriteQueryAnswer, WriteQueryResult,
+    },
     transaction::{TransactionRead, TransactionSchema, TransactionWrite},
 };
 use diagnostics::{
@@ -74,11 +78,21 @@ use crate::service::{
         row::encode_row,
     },
     transaction_service::{
-        execute_schema_query, execute_write_query_in_schema, execute_write_query_in_write, init_transaction_timeout,
-        is_write_pipeline, prepare_read_query_in, unwrap_or_execute_and_return, with_readable_transaction,
-        StreamQueryOutputDescriptor, Transaction, TransactionServiceError, WriteQueryAnswer, WriteQueryResult,
+        init_transaction_timeout, is_write_pipeline, with_readable_transaction, Transaction, TransactionServiceError,
     },
 };
+
+macro_rules! unwrap_or_execute_and_return {
+    ($match_: expr, |$err:pat_param| $err_mapper: block) => {{
+        match $match_ {
+            Ok(inner) => inner,
+            Err($err) => {
+                $err_mapper
+                return;
+            }
+        }
+    }};
+}
 
 #[derive(Debug)]
 pub(crate) struct TransactionService {
@@ -487,7 +501,7 @@ impl TransactionService {
                 Transaction::Schema(transaction)
             }
         };
-        self.diagnostics_manager.increment_load_count(ClientEndpoint::Grpc, &database_name, transaction.to_load_kind());
+        self.diagnostics_manager.increment_load_count(ClientEndpoint::Grpc, &database_name, transaction.load_kind());
         self.transaction = Some(transaction);
         self.timeout_at = init_transaction_timeout(Some(transaction_timeout_millis));
         self.is_open = true;
@@ -837,7 +851,10 @@ impl TransactionService {
         if let Some(transaction) = self.transaction.take() {
             match transaction {
                 Transaction::Schema(schema_transaction) => {
-                    let (transaction, result) = execute_schema_query(schema_transaction, query, source_query).await;
+                    let (transaction, result) =
+                        spawn_blocking(move || execute_schema_query(schema_transaction, query, source_query))
+                            .await
+                            .expect("Expected schema query execution finishing");
                     self.transaction = Some(Transaction::Schema(transaction));
                     let message_ok_done =
                         result.map(|_| query_res_ok_done(typedb_protocol::query::Type::Schema)).map_err(|err| {
@@ -927,10 +944,14 @@ impl TransactionService {
         let interrupt = self.query_interrupt_receiver.clone();
         match self.transaction.take() {
             Some(Transaction::Schema(schema_transaction)) => Ok(spawn_blocking(move || {
-                execute_write_query_in_schema(schema_transaction, query_options, pipeline, source_query, interrupt)
+                let (transaction, result) =
+                    execute_write_query_in_schema(schema_transaction, query_options, pipeline, source_query, interrupt);
+                (Transaction::Schema(transaction), result)
             })),
             Some(Transaction::Write(write_transaction)) => Ok(spawn_blocking(move || {
-                execute_write_query_in_write(write_transaction, query_options, pipeline, source_query, interrupt)
+                let (transaction, result) =
+                    execute_write_query_in_write(write_transaction, query_options, pipeline, source_query, interrupt);
+                (Transaction::Write(transaction), result)
             })),
             Some(Transaction::Read(transaction)) => {
                 self.transaction = Some(Transaction::Read(transaction));
@@ -996,7 +1017,7 @@ impl TransactionService {
         thing_manager: Arc<ThingManager>,
         output_descriptor: StreamQueryOutputDescriptor,
         query_options: QueryOptions,
-        query_structure: Option<QueryStructure>,
+        _query_structure: Option<QueryStructure>,
         batch: Batch,
         sender: Sender<StreamQueryResponse>,
         timeout_at: Instant,
@@ -1124,12 +1145,11 @@ impl TransactionService {
             let query_manager = transaction.query_manager.clone();
             spawn_blocking(move || {
                 let start_time = Instant::now();
-                let pipeline = prepare_read_query_in(
+                let pipeline = query_manager.prepare_read_pipeline(
                     snapshot.clone(),
                     &type_manager,
                     thing_manager.clone(),
                     &function_manager,
-                    &query_manager,
                     &pipeline,
                     &source_query,
                 );
@@ -1334,7 +1354,7 @@ impl TransactionService {
     }
 
     fn get_database_name(&self) -> Option<&str> {
-        self.transaction.as_ref().map(Transaction::get_database_name)
+        self.transaction.as_ref().map(Transaction::database_name)
     }
 }
