+++ server/service/http/transaction_service.rs
@@ -18,6 +18,10 @@ use compiler::query_structure::QueryStructure;
 use concept::{thing::thing_manager::ThingManager, type_::type_manager::TypeManager};
 use database::{
     database_manager::DatabaseManager,
+    query::{
+        execute_schema_query, execute_write_query_in_schema, execute_write_query_in_write, StreamQueryOutputDescriptor,
+        WriteQueryAnswer, WriteQueryResult,
+    },
     transaction::{TransactionRead, TransactionSchema, TransactionWrite},
 };
 use diagnostics::{
@@ -50,9 +54,7 @@ use super::message::query::query_structure::encode_query_structure;
 use crate::service::{
     http::message::query::{document::encode_document, query_structure::QueryStructureResponse, row::encode_row},
     transaction_service::{
-        execute_schema_query, execute_write_query_in_schema, execute_write_query_in_write, init_transaction_timeout,
-        is_write_pipeline, prepare_read_query_in, with_readable_transaction, StreamQueryOutputDescriptor, Transaction,
-        TransactionServiceError, WriteQueryAnswer, WriteQueryResult,
+        init_transaction_timeout, is_write_pipeline, with_readable_transaction, Transaction, TransactionServiceError,
     },
     QueryType, TransactionType,
 };
@@ -281,7 +283,7 @@ impl TransactionService {
                 Transaction::Schema(transaction)
             }
         };
-        self.diagnostics_manager.increment_load_count(ClientEndpoint::Http, &database_name, transaction.to_load_kind());
+        self.diagnostics_manager.increment_load_count(ClientEndpoint::Http, &database_name, transaction.load_kind());
         self.transaction = Some(transaction);
         self.timeout_at = init_transaction_timeout(Some(transaction_timeout_millis));
 
@@ -721,7 +723,10 @@ impl TransactionService {
         if let Some(transaction) = self.transaction.take() {
             match transaction {
                 Transaction::Schema(schema_transaction) => {
-                    let (transaction, result) = execute_schema_query(schema_transaction, query, source_query).await;
+                    let (transaction, result) =
+                        spawn_blocking(move || execute_schema_query(schema_transaction, query, source_query))
+                            .await
+                            .expect("Expected schema query execution finishing");
                     self.transaction = Some(Transaction::Schema(transaction));
                     match result {
                         Ok(_) => return Ok(TransactionServiceResponse::Query(QueryAnswer::ResOk(QueryType::Schema))),
@@ -822,10 +827,14 @@ impl TransactionService {
         let interrupt = self.query_interrupt_receiver.clone();
         match self.transaction.take() {
             Some(Transaction::Schema(schema_transaction)) => Ok(spawn_blocking(move || {
-                execute_write_query_in_schema(schema_transaction, query_options, pipeline, source_query, interrupt)
+                let (transaction, result) =
+                    execute_write_query_in_schema(schema_transaction, query_options, pipeline, source_query, interrupt);
+                (Transaction::Schema(transaction), result)
             })),
             Some(Transaction::Write(write_transaction)) => Ok(spawn_blocking(move || {
-                execute_write_query_in_write(write_transaction, query_options, pipeline, source_query, interrupt)
+                let (transaction, result) =
+                    execute_write_query_in_write(write_transaction, query_options, pipeline, source_query, interrupt);
+                (Transaction::Write(transaction), result)
             })),
             Some(Transaction::Read(transaction)) => {
                 self.transaction = Some(Transaction::Read(transaction));
@@ -974,12 +983,11 @@ impl TransactionService {
             let function_manager = transaction.function_manager.clone();
             let query_manager = transaction.query_manager.clone();
             spawn_blocking(move || {
-                let pipeline_result = prepare_read_query_in(
+                let pipeline_result = query_manager.prepare_read_pipeline(
                     snapshot.clone(),
                     &type_manager,
                     thing_manager.clone(),
                     &function_manager,
-                    &query_manager,
                     &pipeline,
                     &source_query,
                 );
@@ -1162,9 +1170,5 @@ impl TransactionService {
             event!(Level::INFO, "Read query done (including network request time).\n{}", query_profile);
         }
         Continue(())
-    }
-
-    fn get_database_name(&self) -> Option<&str> {
-        self.transaction.as_ref().map(Transaction::get_database_name)
     }
 }
